todo/requirements.txt
alembic==1.13.1
bcrypt==4.1.3
blinker==1.8.2
Brotli==1.1.0
cffi==1.16.0
click==8.1.7
dnspython==2.6.1
email_validator==2.1.2
eventlet==0.36.1
Faker==25.8.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Compress==1.15
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
infinity==1.5
iniconfig==2.0.0
intervals==0.9.2
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
packaging==24.1
pluggy==1.5.0
psycopg2cffi==2.9.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
pytest==8.2.2
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
rich==13.7.1
shellingham==1.5.4
six==1.16.0
SQLAlchemy==2.0.31
SQLAlchemy-Utils==0.41.2
typer==0.12.3
typing_extensions==4.12.2
validators==0.28.3
Werkzeug==3.0.3
WTForms==3.1.2
WTForms-Alchemy==0.18.0
WTForms-Components==0.10.5
zstandard==0.22.0
